(refactor) Move CSS out of App.jsx
