+++ client/src/components/utils/overridingStyleObjects.js
@@ -1,0 +1,27 @@
+const directionsIconStyle = {
+  fill: 'rgb(178,235,242)',
+};
+
+const mapFullScreenStyle = {};
+
+const mapSelectingRouteStyle = {
+  height: 'calc(100% - 112px)',
+};
+
+const textFieldInputStyle = {
+  color: 'rgb(224,247,250)',
+  backgroundColor: 'rgb(77,208,225)',
+  borderRadius: '5px',
+  'box-sizing': 'border-box',
+  height: '80%',
+  paddingLeft: '10px',
+  paddingRight: '10px',
+  width: '100%',
+};
+
+export default {
+  directionsIconStyle,
+  mapFullScreenStyle,
+  mapSelectingRouteStyle,
+  textFieldInputStyle,
+};
